+++ src/main/java/dev/hcr/hcf/factions/commands/FactionCommandManager.java
@@ -1,9 +1,11 @@
 package dev.hcr.hcf.factions.commands;
 
 import dev.hcr.hcf.factions.commands.captain.FactionInviteCommand;
+import dev.hcr.hcf.factions.commands.captain.FactionKickCommand;
 import dev.hcr.hcf.factions.commands.captain.FactionWithdrawCommand;
 import dev.hcr.hcf.factions.commands.coleader.FactionClaimCommand;
 import dev.hcr.hcf.factions.commands.leader.FactionCreateCommand;
+import dev.hcr.hcf.factions.commands.leader.FactionDisbandCommand;
 import dev.hcr.hcf.factions.commands.member.*;
 import dev.hcr.hcf.factions.commands.staff.FactionTestCommand;
 
@@ -19,12 +21,14 @@ public class FactionCommandManager {
         // CAPTAIN COMMANDS
         new FactionInviteCommand();
         new FactionWithdrawCommand();
+        new FactionKickCommand();
 
         // COLEADER COMMANDS
         new FactionClaimCommand();
 
         // LEADER COMMANDS
         new FactionCreateCommand();
+        new FactionDisbandCommand();
 
         // STAFF COMMANDS
         new FactionTestCommand();