+++ src/main/java/dev/hcr/hcf/factions/structure/Relation.java
@@ -1,6 +1,7 @@
 package dev.hcr.hcf.factions.structure;
 
 import dev.hcr.hcf.factions.Faction;
+import dev.hcr.hcf.factions.types.SystemFaction;
 import dev.hcr.hcf.users.User;
 import org.bukkit.ChatColor;
 import org.bukkit.entity.Player;