+++ src/main/java/dev/hcr/hcf/users/User.java
@@ -3,16 +3,22 @@ package dev.hcr.hcf.users;
 import dev.hcr.hcf.factions.Faction;
 import dev.hcr.hcf.factions.types.PlayerFaction;
 import dev.hcr.hcf.users.faction.ChatChannel;
+import dev.hcr.hcf.users.statistics.types.OreStatistics;
+import dev.hcr.hcf.users.statistics.types.PvPStatistics;
 import dev.hcr.hcf.utils.TaskUtils;
+import org.bson.Document;
 import org.bukkit.Bukkit;
 import org.bukkit.entity.Player;
 
+import java.util.Collection;
 import java.util.UUID;
 import java.util.concurrent.ConcurrentHashMap;
 
 public class User {
     private final UUID uuid;
     private final String name;
+    private OreStatistics oreStatistics;
+    private PvPStatistics pvPStatistics;
     private double balance;
     private ChatChannel channel = ChatChannel.PUBLIC;
     private Faction faction;
@@ -22,6 +28,8 @@ public class User {
     public User(UUID uuid, String name) {
         this.uuid = uuid;
         this.name = name;
+        this.pvPStatistics = new PvPStatistics(this);
+        this.oreStatistics = new OreStatistics(this);
         loadFaction();
         users.put(uuid, this);
     }
@@ -32,9 +40,26 @@ public class User {
                 if (!(faction instanceof PlayerFaction)) continue;
                 if (((PlayerFaction) faction).hasMember(uuid)) {
                     this.faction = faction;
+                    return;
                 }
             }
         });
+    }
+
+    public void load(Document document) {
+        this.pvPStatistics = new PvPStatistics(this, document);
+        this.oreStatistics = new OreStatistics(this, document);
+        if (document.containsKey("balance")) {
+            this.balance = document.getDouble("balance");
+        }
+    }
+
+    public Document save() {
+        Document document = new Document("uuid", uuid.toString());
+        document.append("name", this.name);
+        oreStatistics.getStatisticKeyMapping().forEach(document::append);
+        document.append("balance", balance);
+        return document;
     }
 
     public ChatChannel getChannel() {
@@ -83,6 +108,18 @@ public class User {
 
     public Player toPlayer() {
         return Bukkit.getPlayer(uuid);
+    }
+
+    public PvPStatistics getPvPStatistics() {
+        return pvPStatistics;
+    }
+
+    public OreStatistics getOreStatistics() {
+        return oreStatistics;
+    }
+
+    public static Collection<User> getUsers() {
+        return users.values();
     }
 
     public static User getUser(String name) {