+++ src/main/java/dev/hcr/hcf/users/statistics/types/enums/Ores.java
@@ -1,0 +1,11 @@
+package dev.hcr.hcf.users.statistics.types.enums;
+
+public enum Ores {
+    DIAMONDS,
+    EMERALDS,
+    GOLD,
+    LAPIS,
+    REDSTONE,
+    IRON,
+    COAL;
+}